# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Mustashar AI API
  version: 1.0.0
  description: API specification for Mustashar AI - Legal Consultation Platform
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        personaType:
          type: string
          enum:
            - defense-lawyer
            - legal-analyst
            - judge-vision
            - quick-consultation
            - smart-mufti
        messageCount:
          type: integer
        totalCreditsUsed:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - personaType
        - messageCount
        - totalCreditsUsed
        - createdAt
        - updatedAt
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum:
            - user
            - assistant
        content:
          type: string
        creditsUsed:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - creditsUsed
        - createdAt
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
paths:
...
  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      security:
        - BearerAuth: []
      summary: Get all conversations for current user
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      operationId: createConversation
      tags: [conversations]
      security:
        - BearerAuth: []
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                personaType:
                  type: string
                  enum:
                    - defense-lawyer
                    - legal-analyst
                    - judge-vision
                    - quick-consultation
                    - smart-mufti
              required:
                - title
                - personaType
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [conversations]
      security:
        - BearerAuth: []
      summary: Get a conversation with its messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation:
                    $ref: "#/components/schemas/Conversation"
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      operationId: deleteConversation
      tags: [conversations]
      security:
        - BearerAuth: []
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...